refactor: rename celery_app.py to celery.py
Our Ansible playbooks rely on celery_app.py being called celery.py.

https://github.com/openedx/configuration/blob/ec684e35b0e73883be13958c1e2b52f461a0aba7/playbooks/roles/edx_django_service/templates/edx/app/supervisor/conf.d.available/app-workers.conf.j2#L37
