+++ commerce_coordinator/__init__.py
@@ -3,7 +3,7 @@ commerce-coordinator module.
 """
 # This will make sure the Celery app is always imported when Django starts so that the Celery shared_task decorator
 # will use this app.
-from .celery_app import app as celery_app
+from .celery import app as celery_app
 
 __all__ = ('celery_app',)
